+++ packages/config/src/doc/IConfigOpts.ts
@@ -19,9 +19,9 @@ export interface IConfigOpts {
 
     /**
      * Directory where project config files are located. Defaults to working directory.
-     * Specify an empty string to disable loading of project config files.
+     * Specify `false` to disable loading of project config files.
      */
-    projectDir?: string;
+    projectDir?: string | false;
 
     /**
      * Vault object with methods for loading and saving secure credentials